requirements.txt
Flask==2.1.2
opencv-python==4.6.0.66
Flask-SocketIO==5.2.0
simple-websocket==0.5.2
Pillow==9.2.0